Remove a downcast, ints are not long enough to represent anything over 2gb

+++ src/main/java/cd/go/contrib/elasticagents/dockerswarm/elasticagent/DockerService.java
@@ -105,7 +105,7 @@ public class DockerService {
             long memory = Size.parse(request.properties().get("Memory")).toBytes();
             resourceRequirements = ResourceRequirements.builder()
                     .withLimits(resourcesBuilder
-                            .withMemoryBytes((int) memory)
+                            .withMemoryBytes(memory)
                             .build())
                     .build();
         }
